Move upb base entity to separate module
